Add extra config options for recalculating Wait Time
Added 2 new options for recalculating the Wait Time. This allows
the user to specificy whether the Random Timer will recalculate the
Wait Time on `Timeout` or on `start()`. Essentially, this allows
the user to have the timer calculate a random Wait Time once and then
keep using that value moving forward.